+++ Changelog.md
@@ -3,7 +3,7 @@
 This page contains information about changes to the PowerBI Visual Tools (pbiviz).
 
 ## pbiviz v1.10.2
-* Add `pbiviz --create-cert` to generating new unique certificate
+* Add `pbiviz --create-cert` for generating new unique certificate
 
 ## pbiviz v1.10.1
 * Increase typescript to 2.3.3 version